requirements.txt
flask
bibtexparser
tinydb
GitPython
